Use `cf_folds` variable name throughout `estimate_effects`
In the updated version of `estimate_effects`, the input variable has been renamed from `cv_folds` to `cf_folds`, but there are still two uses of the previous name. Calls to this function for me are now throwing: `NameError: name 'cv_folds' is not defined`. 

This simply gives these two uses the new name.